[parser] for now point_size returns integers (reportlab API expect this)

+++ escparser/parser.py
@@ -292,9 +292,9 @@ class ESCParser:
             self.current_pdf.setFillColor(colors.HexColor(self.RGB_colors[self.color]))
 
     @property
-    def point_size(self) -> float:
+    def point_size(self) -> int:
         """Get the current font point size (in cpi)"""
-        return self._point_size
+        return int(self._point_size)
 
     @point_size.setter
     def point_size(self, point_size: float):
